Add config.js
