+++ src/js/model.js
@@ -1,16 +1,16 @@
+import { API_URL } from './config.js';
+import { getJSON } from './helpers.js';
+
 export const state = {
     recipe: {},
 };
 
 export const loadRecipe = async function (id) {
     try {
-        const res = await fetch(`https://forkify-api.herokuapp.com/api/v2/recipes/${id}`);
-        const data = await res.json();
-    
-        if (!res.ok) throw new Error(`${data.message} (${res.status})`);
+        const data = await getJSON(`${API_URL}/${id}`);
     
         const { recipe } = data.data;
-    
+
         state.recipe = {
           id: recipe.id,
           title: recipe.title,
@@ -24,6 +24,7 @@ export const loadRecipe = async function (id) {
     
         console.log(state.recipe);
     } catch (err) {
-        alert(err);
+        // Temp error handling
+        console.error(`${err} 💥💥💥💥`);
     }
 };
